+++ src/interop/antlr4.ts
@@ -1,3 +1,7 @@
+/**
+ * ANTLR4 interoperability module. Requires the antlr4ng implementation.
+ * @module interop/antlr4
+ */
 export * from '../parsing/parse-tree';
 export * from '../parsing/parsing';
 export * from "../mapping";