+++ src/mapping.ts
@@ -1,49 +1,8 @@
-import {ParserRuleContext, ParseTree, TerminalNode} from "antlr4ng";
-import {Child, Node, NODE_DEFINITION_SYMBOL, Origin, registerNodeDefinition} from "./model/model";
-import {ASTTransformer, GenericNode, Mapped, registerNodeFactory, transform} from "./transformation/transformation";
+import {ParserRuleContext, TerminalNode} from "antlr4ng";
+import {Node, Origin} from "./model/model";
+import {ASTTransformer} from "./transformation/transformation";
 import {ParseTreeOrigin} from "./parsing";
 
-/**
- * Registers the decorated node as a target for transformation from the given `type`.
- *
- * Note: this will eventually be integrated with Kolasu-style transformers.
- * @param type the type of the source node to map to this node.
- * @deprecated please use StarLasu AST transformers.
- */
-export function ASTNodeFor<T extends ParseTree>(type: new (...args: any[]) => T) {
-    return function (target: new () => Node): void {
-        if(!target[NODE_DEFINITION_SYMBOL]) {
-            registerNodeDefinition(target);
-        }
-        registerNodeFactory(type, () => new target());
-    };
-}
-
-//-------//
-// toAST //
-//-------//
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export function toAST(tree?: ParseTree | null, parent?: Node): Node | undefined {
-    if (tree == null)
-        tree = undefined;
-
-    const node = transform(tree, parent, toAST);
-    if(node && !node.origin) { //Give a chance to custom factories to set a different node
-        node.origin = new ParseTreeOrigin(tree);
-    }
-    return node;
-}
-
-export class GenericParseTreeNode extends GenericNode {
-    @Child()
-    @Mapped("children")
-    childNodes: GenericParseTreeNode[] = [];
-}
-
-registerNodeFactory(ParserRuleContext, () => new GenericParseTreeNode());
 
 /**
  * Implements a transformation from an ANTLR parse tree (the output of the parser) to an AST (a higher-level