+++ src/transformation/transformation.ts
@@ -1,16 +1,13 @@
 import {
     ASTNode,
-    ensureNodeDefinition,
     getNodeDefinition,
     Node,
-    NODE_DEFINITION_SYMBOL, NodeDefinition,
+    NodeDefinition,
     Origin,
-    registerNodeDefinition,
-    registerNodeProperty
 } from "../model/model";
 import {Issue, IssueSeverity} from "../validation";
 import {Position} from "../model/position";
-import {ErrorNode, GenericErrorNode} from "../model/errors";
+import {GenericErrorNode} from "../model/errors";
 
 export class PropertyRef<Obj, Value> {
     constructor(
@@ -393,197 +390,10 @@ export class ASTTransformer {
     }
 }
 
-//-----------------------------------//
-// Factory and metadata registration //
-//-----------------------------------//
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export const NODE_FACTORY_SYMBOL = Symbol("nodeFactory");
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export const INIT_SYMBOL = Symbol("init");
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export function registerNodeFactory<T>(type: new (...args: any[]) => T, factory: (tree: T) => Node): void {
-    type.prototype[NODE_FACTORY_SYMBOL] = factory;
-}
-
-/**
- * Marks a property of a node as mapped from a property of another node of a different name.
- * @deprecated to be removed, use ParseTreeToASTTranformer
- * @param type the source node's type.
- * @param propertyName the name of the target property.
- * @param path the path in the source node that will be mapped to the target property.
- */
-export function registerPropertyMapping<T extends Node>(
-    type: new (...args: any[]) => T, propertyName: string, path: string = propertyName): any {
-    const propInfo: any = registerNodeProperty(type, propertyName);
-    propInfo.path = path || propertyName;
-    return propInfo;
-}
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export function registerInitializer<T extends Node>(type: new (...args: any[]) => T, methodName: string): void {
-    type[INIT_SYMBOL] = methodName;
-}
-
-//------------//
-// Decorators //
-//------------//
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export function NodeTransform<T extends Node>(type: new (...args: any[]) => T) {
-    return function (target: new () => Node): void {
-        if(!target[NODE_DEFINITION_SYMBOL]) {
-            registerNodeDefinition(target);
-        }
-        registerNodeFactory(type, () => new target());
-    };
-}
-
-/**
- * Marks a property of a node as mapped from a property of another node of a different name.
- * @deprecated to be removed, use ASTTranformer.withChild
- * @param path the path in the source node that will be mapped to the target property.
- */
-export function Mapped(path?: string): (target, methodName: string) => void {
-    return function (target, methodName: string) {
-        registerPropertyMapping(target, methodName, path);
-    };
-}
-
-/**
- * Decorator to register an initializer method on a Node. When a node is instantiated as the target of a
- * transformation, after its properties have been set, the transformer calls the init method, if any.
- * @param target the target type.
- * @param methodName the name of the init method.
- * @deprecated please use StarLasu AST transformers.
- */
-// Since target is any-typed (see https://www.typescriptlang.org/docs/handbook/decorators.html),
-// eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-export function Init(target, methodName: string): void {
-    registerInitializer(target, methodName);
-}
-
-//-----------------//
-// Transformations //
-//-----------------//
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export function fillChildAST<FROM, TO extends Node>(
-    node: TO, property: string, tree: FROM | undefined, transformer: (node: FROM) => TO | undefined): TO[] {
-    const propDef: any = ensureNodeDefinition(node).properties[property];
-    const propertyPath = propDef.path || property;
-    if (propertyPath && propertyPath.length > 0) {
-        const path = propertyPath.split(".");
-        let error;
-        for (const segment in path) {
-            if (tree && (typeof(tree[path[segment]]) === "function")) {
-                try {
-                    tree = tree[path[segment]]();
-                } catch (e) {
-                    error = e;
-                    break;
-                }
-            } else if (tree && tree[path[segment]]) {
-                tree = tree[path[segment]];
-            } else {
-                tree = undefined;
-                break;
-            }
-        }
-        if(error) {
-            node[property] = new GenericErrorNode(error);
-        } else if (tree) {
-            if(propDef.child) {
-                if (Array.isArray(tree)) {
-                    node[property] = [];
-                    for (const i in tree) {
-                        node[property].push(transformer(tree[i])?.withParent(node));
-                    }
-                    return node[property];
-                } else {
-                    node[property] = transformer(tree)?.withParent(node);
-                    return [node[property]];
-                }
-            } else {
-                node[property] = tree;
-            }
-        }
-    }
-    return [];
-}
-
-function makeNode(factory, tree: unknown) {
-    try {
-        return factory(tree) as Node;
-    } catch (e) {
-        return new GenericErrorNode(e);
-    }
-}
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export function transform(tree: unknown, parent?: Node, transformer: typeof transform = transform): Node | undefined {
-    if (typeof tree !== "object" || !tree) {
-        return undefined;
-    }
-    const factory = tree[NODE_FACTORY_SYMBOL];
-    let node: Node;
-    if (factory) {
-        node = makeNode(factory, tree);
-        const def = getNodeDefinition(node);
-        if (def) {
-            for (const p in def.properties) {
-                fillChildAST(node, p, tree, transformer);
-            }
-        }
-        const initFunction = node[INIT_SYMBOL];
-        if (initFunction) {
-            try {
-                node[initFunction].call(node, tree);
-            } catch (e) {
-                node = new PartiallyInitializedNode(node, e);
-            }
-        }
-    } else {
-        node = new GenericNode();
-    }
-    return node.withParent(parent);
-}
-
 @ASTNode("com.strumenta.tylasu.transformation", "GenericNode")
 export class GenericNode extends Node {
     constructor(parent?: Node) {
         super();
         this.parent = parent;
-    }
-}
-
-/**
- * @deprecated please use StarLasu AST transformers.
- */
-export class PartiallyInitializedNode extends Node implements ErrorNode {
-    message: string;
-
-    get position(): Position | undefined {
-        return this.node.position;
-    }
-
-    constructor(readonly node: Node, error: Error) {
-        super();
-        this.message = `Could not initialize node: ${error}`;
     }
 }